[#86bw58pwd] Fixed bullet movement

+++ src/main/scala/org/aas/sbtanks/entities/bullet/Bullet.scala
@@ -1,3 +1,3 @@
 package org.aas.sbtanks.entities.bullet
 
-case class Bullet(speed: Int, isPlayerBullet: Boolean)
+case class Bullet(speed: Double, isPlayerBullet: Boolean)
